Added routes and fixed formatting
Fixed formatting issues of club page where the join-button-container overlaps with navbar and makes it unable to click the login and logout button by raising navbars z-index. Also added routes to delete all of a users applications to either member or executive and implemented those routes after approval or rejection.